+++ yandex/resource_yandex_compute_instance_group.go
@@ -270,7 +270,7 @@ func resourceYandexComputeInstanceGroup() *schema.Resource {
 						"platform_id": {
 							Type:     schema.TypeString,
 							Optional: true,
-							Default:  "standard-v1",
+							Default:  "standard-v2",
 						},
 
 						"description": {